+++ app/src/types/pdf-lib-fontkit.d.ts
@@ -1,0 +1,5 @@
+declare module '@pdf-lib/fontkit' {
+    import type { Fontkit } from 'pdf-lib-plus-encrypt';
+    const fontkit: Fontkit;
+    export default fontkit;
+}